Backend/requirements.txt
streamlit==1.28.1
supabase==2.3.0
python-dotenv==1.0.0